source/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.1
Werkzeug==0.9.6
certifi==14.05.14
django-appconf==0.6
django-classy-tags==0.5.1
django-compressor==1.4
django-debug-toolbar==1.2.1
django-extensions==1.3.8
django-model-utils==2.2
django-sekizai==0.7
gunicorn==18.0
ipdb==0.8
ipython==2.1.0
# mysql-connector-python==1.2.2
psycopg2==2.5.2
pyparsing==2.0.2
pyzmq==14.3.1
requests==2.3.0
selenium==2.42.1
setproctitle==1.1.8
six==1.7.3
sqlparse==0.1.11
tornado==4.0
